Use accessors in RofiView to get modi's and num modi for sidebar.

+++ include/view-internal.h
@@ -63,7 +63,7 @@ typedef struct RofiViewState
     workarea     mon;
 
     // Sidebar view
-    ssize_t      num_modi;
+    unsigned int num_modi;
     textbox      **modi;
 
     MenuFlags    menu_flags;
